+++ assets/scripts/ziti.js
@@ -19,6 +19,10 @@ $(document).ready(function(e) {
 var app = {
 	keys: [],
 	isDirty: false,
+	identityRoles: null,
+	hostedRoles: null,
+	idRoles: null,
+	typedIn: false,
 	init: function() {
 		app.events();
 		app.setupLock();
@@ -34,6 +38,22 @@ var app = {
 		if (settings) settings.init();
 		if (restrictions) restrictions.init();
 		$('*[data-go="'+window.location.pathname+'"]').addClass("selected");
+
+		app.identityRoles = new MultiSelect("IdRoles", 10, true);
+		app.identityRoles.addSource(new SelectSource("identity-role-attributes", "", "id")); 
+		app.identityRoles.init();
+
+		app.hostedRoles = new MultiSelect("WhereHosted", 10, true);
+		app.hostedRoles.appendHash = true;
+		app.hostedRoles.addSource(new SelectSource("identities", "@", "name", "role")); 
+		app.hostedRoles.addSource(new SelectSource("identity-role-attributes", "#", "id")); 
+		app.hostedRoles.init();
+
+		app.idRoles = new MultiSelect("WhoAccesses", 10, true);
+		app.idRoles.appendHash = true;
+		app.idRoles.addSource(new SelectSource("identities", "@", "name", "role")); 
+		app.idRoles.addSource(new SelectSource("identity-role-attributes", "#", "id")); 
+		app.idRoles.init();
  	},
 	events: function() {
 		if (localStorage.getItem("hideTags")=="yes") $("#TagArea").addClass("hidden");
@@ -48,6 +68,124 @@ var app = {
 		$(".modal").find("input").change((e) => { app.isDirty = true; });
 		context.addListener(settings.name, app.settingsReturned);
 		context.addListener("version", app.versionReturned);
+		$("#SServiceName").keyup(app.name);
+		$("#SServiceHost").keyup(app.typed);
+		$("#CreateButton").click(app.createSService);
+		$("#CreateIdButton").click(app.createId);
+		$("#IdentityDownload").click(app.download);
+		$("#DoneIdButton").click(app.idReset);
+		$("#DoneServiceButton").click(app.idReset);
+		$("#InlineAddIdentityButton").click(app.showInlineId);
+		$("#InlineAddServiceButton").click(app.showInlineService);
+	},
+	showInlineService: function() {
+		$("#Select1").hide();
+		$("#InlineServiceArea").show();
+	},
+	showInlineId: function() {
+		$("#Select1").hide();
+		$("#InlineIdentityArea").show();
+	},
+	idReset: function(e) {
+		app.identityRoles.val([]);
+		$("#IdName").val("");
+		$("#Identity2").hide();
+		$("#Identity1").hide();
+		$("#Select1").show();
+		modal.close();
+	},
+	createSService: function(e) {
+		$("#Service1").hide();
+		$("#SServiceLoader").show();
+		var params = {
+			name: $("#SServiceName").val(),
+			protocol: $("#HostedProtocol").val(),
+			host: $("#HostedHost").val(),
+			port: $("#HostedPort").val(),
+			encrypt: $("#SimpleEncryptionRequired").hasClass("on"),
+			zitiHost: $("#HowTo").val(),
+			zitiPort: $("#HowToPort").val(),
+			hosted: app.hostedRoles.val(),
+			access: app.idRoles.val(),
+		};
+		service.call("service", params, (e) => {
+			$("#SServiceLoader").hide();
+			if (e.error) {
+				growler.error(e.error);
+				$("#Service1").show();
+			} else {
+				let html = "";
+				let lastTitle = "";
+				for (let i=0; i<e.data.length; i++) {
+					let log = e.data[i];
+					if (log.type!=lastTitle) {
+						html += '<label>'+log.type+'</label>';
+						lastTitle = log.type;
+					}
+					html += '<div class="grid split"><div>'+log.name+'</div><div>'+log.id+'</div></div>';
+				}
+				$("#ServiceHappen").html(html);
+				$("#Service2").show();
+			}
+		});
+
+	},
+	createId: function(e) {
+		$("#IdentityLoader").show();
+		$("#Identity1").hide();
+		var params = {
+			name: $("#IdName").val(),
+			roles: app.identityRoles.val()
+		}
+		service.call("identity", params, (e) => {
+			$("#IdentityLoader").hide();
+			if (e.error) {
+				growler.error(e.error);
+				$("#Identity1").show();
+			} else {
+				app.lastId = e.data;
+				$("#Identity2").show();
+				app.genQR();
+			}
+		});
+	},
+	download: function(e) {
+		var identity = app.lastId;
+		var jwt = "";
+		if (identity&&identity.enrollment&&identity.enrollment.ott&&identity.enrollment.ott.jwt) jwt = identity.enrollment.ott.jwt;
+		else jwt = identity.enrollment.updb.jwt;
+		var name = identity.name.split(" ").join("");
+		var element = document.createElement('a');
+		element.setAttribute('href', 'data:application/ziti-jwt;charset=utf-8,' + encodeURIComponent(jwt));
+		element.setAttribute('download', name+".jwt");
+		element.style.display = 'none';
+		document.body.appendChild(element);
+		element.click();
+		document.body.removeChild(element);
+	},
+	genQR: function(e) {
+		var identity = app.lastId;
+		var jwt = "";
+		if (identity&&identity.enrollment&&identity.enrollment.ott&&identity.enrollment.ott.jwt) jwt = identity.enrollment.ott.jwt;
+		else jwt = identity.enrollment.updb.jwt;
+		$("#QRCode").html("");
+		new QRCode("IdentityQrCode", {
+			text: encodeURIComponent(jwt),
+			width: 300,
+			height: 300,
+			colorDark : "#000000",
+			colorLight : "#ffffff",
+			correctLevel : QRCode.CorrectLevel.M
+		});
+	},
+	typed: function(e) {
+		app.typedIn = true;
+	},
+	name: function(e) {
+		if (!app.typedIn) {
+			var currentUrl = $("#SServiceName").val().trim().split(' ').join('').replace(/[^a-z0-9]/gi, '').toLowerCase();
+			$("#HowTo").val(currentUrl+".ziti");
+		}
 	},
 	keypress: function(e) {
 		if (e.keyCode === 27) {